maps/ctmap: Create NAT maps with correct BPF map type
NAT maps were previously always create as type MapTypeLRUHash.  However,
to use these Go objects to create the BPF maps on the kernel side, we
need the Go and BPF map types to be the same.

Because of this change, we also cannot initialize ctmap's mapInfo in
ctmap.init() and need to use a singleton (sync.Once) pattern.  We also
don't need to call ctmap.InitMapInfo from NewDaemon since that
information will be lazily initialized when it's used.  Thus,
ctmap.InitMapInfo is unexported.

Signed-off-by: Paul Chaignon <paul@cilium.io>